+++ lib/multi_state_sheet/render_multi_state_sheet.dart
@@ -52,6 +52,9 @@ class _MultiStateSheetWidget<StateType>
   /// Determines if the content should be kept behind the footer and not cut by it.
   final bool keepContentBehindFooter;
 
+  /// Padding for the bottom part of safe area around the sheet.
+  final double? safeAreaBottomPadding;
+
   /// Whether to resize the sheet to avoid bottom padding (e.g., keyboard).
   final bool resizeToAvoidBottomPadding;
 
@@ -71,6 +74,7 @@ class _MultiStateSheetWidget<StateType>
     this.drawOutsideWidgetBehindBackgroundFill = false,
     this.offsetOutsideWidgetByTopheader = true,
     this.keepContentBehindFooter = false,
+    this.safeAreaBottomPadding,
     this.resizeToAvoidBottomPadding = true,
     super.key,
   });
@@ -97,6 +101,7 @@ class _MultiStateSheetWidget<StateType>
             barrierColorDelegate: barrierColorDelegate,
             backgroundColor: backgroundColor,
             safeAreaColor: safeAreaColor,
+            safeAreaBottomPadding: safeAreaBottomPadding,
             shaper: shaper,
             drawOutsideWidgetBehindBackgroundFill:
                 drawOutsideWidgetBehindBackgroundFill,
@@ -115,6 +120,7 @@ class _MultiStateSheetWidget<StateType>
       ..barrierColorDelegate = barrierColorDelegate
       ..backgroundColor = backgroundColor
       ..safeAreaColor = safeAreaColor
+      ..safeAreaBottomPadding = safeAreaBottomPadding
       ..shaper = shaper
       ..drawOutsideWidgetBehindBackgroundFill =
           drawOutsideWidgetBehindBackgroundFill
@@ -133,23 +139,125 @@ class _MultiStateSheetWidget<StateType>
 /// the height of the sheet, and adjusting for safe area padding.
 class _RenderMultiStateSheet<StateType> extends RenderBox
     with SlottedContainerRenderObjectMixin<_MultiStateSheetSlot, RenderBox> {
-  MultiStateSheetController<StateType> scrollController;
-  StatefulSheetDelegate<Color?>? barrierColorDelegate;
-  StatefulSheetDelegate<double>? outsideOpacityDelegate;
-  StatefulSheetDelegate<double>? topHeaderOffset;
-  Color backgroundColor;
-  Color? safeAreaColor;
-  ShapeBorder? shaper;
-  bool drawOutsideWidgetBehindBackgroundFill;
-  bool offsetOutsideWidgetByTopheader;
-  bool keepContentBehindFooter;
-  double? topHeaderOffsetValue;
+  MultiStateSheetController<StateType> _scrollController;
+  MultiStateSheetController<StateType> get scrollController =>
+      _scrollController;
+  set scrollController(MultiStateSheetController<StateType> value) {
+    if (_scrollController != value) {
+      _scrollController.removeListener(onSheetOffsetChanges);
+      _scrollController = value;
+      _scrollController.addListener(onSheetOffsetChanges);
+      markNeedsLayout();
+    }
+  }
+
+  StatefulSheetDelegate<Color?>? _barrierColorDelegate;
+  StatefulSheetDelegate<Color?>? get barrierColorDelegate =>
+      _barrierColorDelegate;
+  set barrierColorDelegate(StatefulSheetDelegate<Color?>? value) {
+    if (_barrierColorDelegate != value) {
+      _barrierColorDelegate = value;
+      markNeedsPaint();
+    }
+  }
+
+  StatefulSheetDelegate<double>? _outsideOpacityDelegate;
+  StatefulSheetDelegate<double>? get outsideOpacityDelegate =>
+      _outsideOpacityDelegate;
+  set outsideOpacityDelegate(StatefulSheetDelegate<double>? value) {
+    if (_outsideOpacityDelegate != value) {
+      _outsideOpacityDelegate = value;
+      markNeedsLayout();
+    }
+  }
+
+  StatefulSheetDelegate<double>? _topHeaderOffset;
+  StatefulSheetDelegate<double>? get topHeaderOffset => _topHeaderOffset;
+  set topHeaderOffset(StatefulSheetDelegate<double>? value) {
+    if (_topHeaderOffset != value) {
+      _topHeaderOffset = value;
+      markNeedsLayout();
+    }
+  }
+
+  Color _backgroundColor;
+  Color get backgroundColor => _backgroundColor;
+  set backgroundColor(Color value) {
+    if (_backgroundColor != value) {
+      _backgroundColor = value;
+      markNeedsPaint();
+    }
+  }
+
+  Color? _safeAreaColor;
+  Color? get safeAreaColor => _safeAreaColor;
+  set safeAreaColor(Color? value) {
+    if (_safeAreaColor != value) {
+      _safeAreaColor = value;
+      markNeedsPaint();
+    }
+  }
+
+  ShapeBorder? _shaper;
+  ShapeBorder? get shaper => _shaper;
+  set shaper(ShapeBorder? value) {
+    if (_shaper != value) {
+      _shaper = value;
+      markNeedsPaint();
+    }
+  }
+
+  bool _drawOutsideWidgetBehindBackgroundFill;
+  bool get drawOutsideWidgetBehindBackgroundFill =>
+      _drawOutsideWidgetBehindBackgroundFill;
+  set drawOutsideWidgetBehindBackgroundFill(bool value) {
+    if (_drawOutsideWidgetBehindBackgroundFill != value) {
+      _drawOutsideWidgetBehindBackgroundFill = value;
+      markNeedsPaint();
+    }
+  }
+
+  bool _offsetOutsideWidgetByTopheader;
+  bool get offsetOutsideWidgetByTopheader => _offsetOutsideWidgetByTopheader;
+  set offsetOutsideWidgetByTopheader(bool value) {
+    if (_offsetOutsideWidgetByTopheader != value) {
+      _offsetOutsideWidgetByTopheader = value;
+      markNeedsLayout();
+    }
+  }
+
+  bool _keepContentBehindFooter;
+  bool get keepContentBehindFooter => _keepContentBehindFooter;
+  set keepContentBehindFooter(bool value) {
+    if (_keepContentBehindFooter != value) {
+      _keepContentBehindFooter = value;
+      markNeedsLayout();
+    }
+  }
+
+  double? _topHeaderOffsetValue;
+  double? get topHeaderOffsetValue => _topHeaderOffsetValue;
+  set topHeaderOffsetValue(double? value) {
+    if (_topHeaderOffsetValue != value) {
+      _topHeaderOffsetValue = value;
+      markNeedsLayout();
+    }
+  }
 
   double? _viewBottomPadding;
   double get viewBottomPadding => _viewBottomPadding ?? 0.0;
   set viewBottomPadding(double? value) {
     if (_viewBottomPadding != value) {
       _viewBottomPadding = value;
+      markNeedsLayout();
+    }
+  }
+
+  double? _safeAreaBottomPadding;
+  double get safeAreaBottomPadding => _safeAreaBottomPadding ?? 0.0;
+  set safeAreaBottomPadding(double? value) {
+    if (_safeAreaBottomPadding != value) {
+      _safeAreaBottomPadding = value;
       markNeedsLayout();
     }
   }
@@ -174,20 +282,31 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
   int outsideOpacity = 255;
 
   _RenderMultiStateSheet({
-    required this.scrollController,
-    required this.shaper,
-    required this.backgroundColor,
-    this.topHeaderOffset,
-    this.safeAreaColor,
-    this.barrierColorDelegate,
-    this.outsideOpacityDelegate,
-    this.drawOutsideWidgetBehindBackgroundFill = false,
-    this.offsetOutsideWidgetByTopheader = true,
-    this.keepContentBehindFooter = false,
+    required MultiStateSheetController<StateType> scrollController,
+    required ShapeBorder? shaper,
+    required Color backgroundColor,
+    StatefulSheetDelegate<double>? topHeaderOffset,
+    Color? safeAreaColor,
+    double? safeAreaBottomPadding,
+    StatefulSheetDelegate<Color?>? barrierColorDelegate,
+    StatefulSheetDelegate<double>? outsideOpacityDelegate,
+    bool drawOutsideWidgetBehindBackgroundFill = false,
+    bool offsetOutsideWidgetByTopheader = true,
+    bool keepContentBehindFooter = false,
     bool resizeToAvoidBottomPadding = true,
-    double? viewBottomPadding,
-  })  : draggedSheetOffset = scrollController._extent.offset,
-        _viewBottomPadding = viewBottomPadding,
+  })  : _scrollController = scrollController,
+        _barrierColorDelegate = barrierColorDelegate,
+        _outsideOpacityDelegate = outsideOpacityDelegate,
+        _topHeaderOffset = topHeaderOffset,
+        _backgroundColor = backgroundColor,
+        _safeAreaColor = safeAreaColor,
+        _safeAreaBottomPadding = safeAreaBottomPadding,
+        _shaper = shaper,
+        _drawOutsideWidgetBehindBackgroundFill =
+            drawOutsideWidgetBehindBackgroundFill,
+        _offsetOutsideWidgetByTopheader = offsetOutsideWidgetByTopheader,
+        _keepContentBehindFooter = keepContentBehindFooter,
+        draggedSheetOffset = scrollController._extent.offset,
         _resizeToAvoidBottomPadding = resizeToAvoidBottomPadding;
 
   // Slot getters for accessing child render objects.
@@ -198,7 +317,10 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
   RenderBox? get outside => childForSlot(_MultiStateSheetSlot.outside);
 
   double get sheetHeightExtent => math.max(
-      0.0, constraints.maxHeight - draggedSheetOffset - viewBottomPadding);
+      0.0,
+      constraints.maxHeight -
+          draggedSheetOffset -
+          math.max(0.0, viewBottomPadding - safeAreaBottomPadding));
 
   @override
   bool get sizedByParent => true;
@@ -318,7 +440,8 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
     // Calculate available space for the bottom sheet.
     final constraints = this.constraints;
 
-    final avaliableHeight = constraints.maxHeight - viewBottomPadding;
+    final avaliableHeight = constraints.maxHeight -
+        math.max(0.0, viewBottomPadding - safeAreaBottomPadding);
 
     /// Updates the `lastMaxOffset` of the scroll controller's extent behavior to the current `maxOffset`.
     ///
@@ -482,6 +605,14 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
       _positionChild(outside, constraints, kStartOfTheViewport);
     }
 
+    print([
+      avaliableHeight,
+      draggedSheetOffset.ceil(),
+      scrollController._extent.behavior.snappingPixelOffsetsClamped,
+      scrollController._extent.minOffset,
+      scrollController._extent.maxOffset,
+    ]);
+
     /// Setup children's offsets
     _positionChild(
         content, constraints, offsetSheetFromTop + headerLayoutExtend);
@@ -522,7 +653,8 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
             kStartOfTheViewport,
             draggedSheetOffset,
             constraints.maxWidth,
-            sheetHeightExtent + viewBottomPadding,
+            sheetHeightExtent +
+                math.max(0.0, viewBottomPadding - safeAreaBottomPadding),
           ),
         ) ??
         (Path()
@@ -532,7 +664,8 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
                 kStartOfTheViewport,
                 draggedSheetOffset,
                 constraints.maxWidth,
-                sheetHeightExtent + viewBottomPadding,
+                sheetHeightExtent +
+                    math.max(0.0, viewBottomPadding - safeAreaBottomPadding),
               ),
               topLeft: const Radius.circular(kDefaultRadius),
               topRight: const Radius.circular(kDefaultRadius),
@@ -597,15 +730,14 @@ class _RenderMultiStateSheet<StateType> extends RenderBox
       Rect.fromLTWH(
         offset.dx,
         constraints.maxHeight -
-            viewBottomPadding +
+            math.max(0.0, viewBottomPadding - safeAreaBottomPadding) +
             offset.dy -
             safeAreaCorrection +
-            kBottomNavigationBarHeight *
-                (safeAreaAnimationController?.value ?? 0.0),
+            safeAreaBottomPadding * (safeAreaAnimationController?.value ?? 0.0),
         constraints.maxWidth,
-        viewBottomPadding > 0.0
+        math.max(0.0, viewBottomPadding - safeAreaBottomPadding) > 0.0
             ? 0.0
-            : kBottomNavigationBarHeight + safeAreaCorrection,
+            : safeAreaBottomPadding + safeAreaCorrection,
       ),
       painter,
     );